Fix Style/OneClassPerFile offenses
This does the following:

- Remove obsolete Gem monkey-patch
- Move supporting overrides for GObject-based models to modules in
  separate files
- Move UI-related overrides for Alexandria models to modules in separate
  files
- Move remaining method on Alexandria::UI to that module's main file
