+++ lib/alexandria/ui/gtk_action_group_overrides.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+# This file is part of Alexandria.
+#
+# See the file README.md for authorship and licensing information.
+
+module Alexandria
+  module UI
+    module GtkActionGroupOverrides
+      def [](index)
+        get_action(index)
+      end
+    end
+
+    Gtk::ActionGroup.prepend GtkActionGroupOverrides
+  end
+end
