+++ lib/alexandria/ui/main_app.rb
@@ -34,7 +34,6 @@ require "alexandria/ui/smart_library_properties_dialog"
 require "alexandria/ui/new_smart_library_dialog"
 require "alexandria/ui/bad_isbns_dialog"
 require "alexandria/ui/dndable"
-require "alexandria/ui/init"
 require "alexandria/ui/ui_manager"
 require "alexandria/ui/listview"
 require "alexandria/ui/icon_view_manager"
